🚚 Move other componenets to page components
