+++ src/lib/PageComponents/Header/LeaderboardFooter.svelte
